correct

+++ app/FinanceX/callbacks/edit_callbacks.py
@@ -151,7 +151,7 @@ def calculate_saldo(n_clicks, saldo_input_data, transaction_data):
     # Convert transaction data to DataFrame
     transactions_df = pd.DataFrame(transaction_data)
     transactions_df['Buchungsdatum'] = pd.to_datetime(transactions_df['Buchungsdatum'])
-    transactions_df['Betrag']=float(transactions_df['Betrag'])
+    transactions_df['Betrag']=transactions_df['Betrag'].astype(float)
     
     # Initialize list to hold calculated saldo for each account
     calculated_saldos = []
